EFK-FE-19 refactor: move enums in it's own file
