+++ efk-fe/src/redux/thunks/categoryThunk.ts
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { ENDPOINT } from '../../constants';
+import { ENDPOINT } from '../../enums';
 import { CategoryDataToSubmit, CategoryWithWords } from '../../interfaces';
 import { clientAPI } from '../../lib';
