+++ efk-fe/src/utils/isGameModePlay.ts
@@ -1,3 +1,3 @@
-import { GAME_MODE } from '../constants';
+import { GAME_MODE } from '../enums';
 
 export const isGameModePlay = (gameMode: GAME_MODE): boolean => gameMode === GAME_MODE.PLAY;
