+++ efk-fe/src/utils/isGameModeTrain.ts
@@ -1,3 +1,3 @@
-import { GAME_MODE } from '../constants';
+import { GAME_MODE } from '../enums';
 
 export const isGameModeTrain = (gameMode: GAME_MODE): boolean => gameMode === GAME_MODE.TRAIN;
