+++ efk-fe/src/utils/isGameStarted.ts
@@ -1,3 +1,3 @@
-import { GAME_STATUS } from '../constants';
+import { GAME_STATUS } from '../enums';
 
 export const isGameStarted = (gameStatus: GAME_STATUS): boolean => gameStatus === GAME_STATUS.STARTED;
