+++ efk-fe/src/utils/isImageType.ts
@@ -1,3 +1,3 @@
-import { FILE_TYPE } from '../constants';
+import { FILE_TYPE } from '../enums';
 
 export const isImageType = (fileType: FILE_TYPE): boolean => fileType === FILE_TYPE.IMAGE;
