+++ efk-fe/src/utils/isRoleAdmin.ts
@@ -1,4 +1,4 @@
-import { ROLE } from '../constants';
+import { ROLE } from '../enums';
 
 export const isRoleAdmin = (role: ROLE): boolean => {
   return role === ROLE.ADMIN;
